split lsp limits config model into microcrate
