+++ Cargo.toml
@@ -45,6 +45,7 @@ members = [
     "crates/perl-lsp-completion",
     "crates/perl-lsp-code-actions",
     "crates/perl-lsp-cancellation",
+    "crates/perl-lsp-limits-types",
     "crates/perl-lsp-limits",
     "crates/perl-lsp-feature-policy",
     "crates/perl-diagnostics-codes",
@@ -184,6 +185,7 @@ allow = [
     "perl-lsp-feature-grid",
     "perl-lsp-feature-governance",
     "perl-lsp-launcher",
+    "perl-lsp-limits-types",
     "perl-lsp-limits",
     "perl-lsp-protocol",
     "perl-lsp-transport",
@@ -284,6 +286,7 @@ perl-lsp-semantic-tokens = { path = "crates/perl-lsp-semantic-tokens", version =
 perl-lsp-inlay-hints = { path = "crates/perl-lsp-inlay-hints", version = "0.10.0" }
 perl-lsp-feature-contracts = { path = "crates/perl-lsp-feature-contracts", version = "0.10.0" }
 perl-lsp-cancellation = { path = "crates/perl-lsp-cancellation", version = "0.10.0" }
+perl-lsp-limits-types = { path = "crates/perl-lsp-limits-types", version = "0.10.0" }
 perl-lsp-limits = { path = "crates/perl-lsp-limits", version = "0.10.0" }
 perl-lsp-feature-grid = { path = "crates/perl-lsp-feature-grid", version = "0.10.0" }
 perl-lsp-feature-profile = { path = "crates/perl-lsp-feature-profile", version = "0.10.0" }
