+++ crates/perl-lsp-limits-types/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "perl-lsp-limits"
+name = "perl-lsp-limits-types"
 version.workspace = true
 edition.workspace = true
 rust-version.workspace = true
@@ -7,11 +7,11 @@ authors.workspace = true
 license.workspace = true
 repository.workspace = true
 homepage.workspace = true
-description = "SRP microcrate for bounded LSP limits runtime state"
+description = "SRP microcrate for LSP limits configuration shapes"
 readme = "README.md"
 
 [dependencies]
-perl-lsp-limits-types = { workspace = true }
+serde_json = "1.0.149"
 
 [lints]
 workspace = true
