+++ crates/perl-lsp-limits-types/README.md
@@ -1,0 +1,7 @@
+# perl-lsp-limits-types
+
+Configuration types and profile presets for bounded LSP limits.
+
+This crate intentionally focuses on the `LspLimits` data model and settings
+projection logic so runtime/global state management can live in a separate
+crate.
