+++ crates/perl-lsp-limits/README.md
@@ -1,14 +1,14 @@
 # perl-lsp-limits
 
-Standalone microcrate for LSP operation limits and deadline policy.
+Runtime-state microcrate for LSP operation limits and deadline policy.
 
 ## Scope
 
 This crate owns:
 
-- Result caps (`workspace_symbol_cap`, `references_cap`, etc.)
-- Deadlines for bounded request latency
-- Cache and indexing guardrails
-- Global, thread-safe limit state and helper accessors
+- Global, thread-safe limit state (`LSP_LIMITS`)
+- Fast helper accessors (`workspace_symbol_cap`, `references_cap`, etc.)
+- Re-export of the shared `LspLimits` configuration type
 
-Keeping these concerns in one place supports SRP and allows other crates to share the same bounded behavior contract.
+The concrete limits configuration shape and profile/update logic live in
+`perl-lsp-limits-types` to preserve SRP boundaries.
